Replace sqlite with postgres
- Rename Tenancy Service classes to better reflect namespace
- Create area_definitions table to store them locally
- Update scripts
