+++ .rubocop.yml
@@ -7,9 +7,6 @@ AllCops:
   Exclude:
     # Automatically Generated by rails, so assuming correct
     - db/schema.rb
-    - db/cable_schema.rb
-    - db/cache_schema.rb
-    - db/queue_schema.rb
     - db/migrate/**
     - bin/**/*
     - storage/**/*
