+++ app/models/area_definitions.rb
@@ -1,0 +1,5 @@
+# Area Definition Table
+class AreaDefinition < ApplicationRecord
+  validates :label, presence: true
+  validates :code, presence: true
+end
