+++ db/migrate/20250920222508_create_area_definitions_table.rb
@@ -1,0 +1,9 @@
+class CreateAreaDefinitionsTable < ActiveRecord::Migration[8.0]
+  def change
+    create_table :area_definitions do |t|
+      t.string :label
+      t.string :code
+      t.timestamps
+    end
+  end
+end
